refactor: consolidate to src/server/* delegating to core; single LocalSendServer
